Exercises for margin-and-padding completed
The first exercise was done exactly as intended
after reading the self-check. Beforehand, it
looked sort of similar but not everything was
perfect.
The second exercise has some differences even
after checking the solution but the solution
looks the same. The self-check in both scenarios
was extremely helpful to iron out the details.